Update l2_interfaces.py
Updated to scan configuration of port for "switchport mode" and if access, puts in access bits, if trunk puts in trunk bits.  If neither looks for other lines and potentially adds both (new).
Also can identify a configured access port without a vlan explicitly assigned and put it in vlan 1 (as cisco does).